Combine mpp_port with internal_port
