+++ server/src/assembly/resources/conf/iotdb-engine.properties
@@ -45,9 +45,6 @@ rpc_port=6667
 # Datatype: int
 # data_block_manager_keep_alive_time_in_ms=1000
 
-# Datatype: int
-# mpp_port=7777
-
 # Datatype: String
 # used for communication between cluster nodes.
 # if this parameter is commented, then the IP that binded by the hostname will be used.
