+++ server/src/main/java/org/apache/iotdb/db/conf/IoTDBConfig.java
@@ -109,9 +109,6 @@ public class IoTDBConfig {
   /** Port which the JDBC server listens to. */
   private int rpcPort = 6667;
 
-  /** Port which is used for node communication in MPP. */
-  private int mppPort = 7777;
-
   /** Port which the influxdb protocol server listens to. */
   private int influxDBRpcPort = 8086;
 
@@ -2738,14 +2735,6 @@ public class IoTDBConfig {
 
   public void setSeriesPartitionSlotNum(int seriesPartitionSlotNum) {
     this.seriesPartitionSlotNum = seriesPartitionSlotNum;
-  }
-
-  public int getMppPort() {
-    return mppPort;
-  }
-
-  public void setMppPort(int mppPort) {
-    this.mppPort = mppPort;
   }
 
   public int getDataBlockManagerPort() {
