+++ server/src/main/java/org/apache/iotdb/db/mpp/execution/scheduler/SimpleFragInstanceDispatcher.java
@@ -48,11 +48,12 @@ public class SimpleFragInstanceDispatcher implements IFragInstanceDispatcher {
           TSendFragmentInstanceResp resp = new TSendFragmentInstanceResp(false);
           try {
             for (FragmentInstance instance : instances) {
+              // TODO: (jackie tien) change the port
               InternalService.Iface client =
-                  InternalServiceClientFactory.getMppServiceClient(
+                  InternalServiceClientFactory.getInternalServiceClient(
                       new Endpoint(
                           instance.getHostEndpoint().getIp(),
-                          IoTDBDescriptor.getInstance().getConfig().getMppPort()));
+                          IoTDBDescriptor.getInstance().getConfig().getInternalPort()));
               // TODO: (xingtanzjr) consider how to handle the buffer here
               ByteBuffer buffer = ByteBuffer.allocate(1024 * 1024);
               instance.serializeRequest(buffer);
