+++ thrift/src/main/thrift/mpp.thrift
@@ -16,9 +16,31 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
+include "common.thrift"
 namespace java org.apache.iotdb.mpp.rpc.thrift
 
+struct CreateSchemaRegionReq {
+    1: required common.TRegionReplicaSet regionReplicaSet
+    2: required string storageGroup
+}
+
+struct CreateDataRegionReq {
+    1: required common.TRegionReplicaSet regionReplicaSet
+    2: required string storageGroup
+    3: optional i64 ttl
+}
+
+struct MigrateSchemaRegionReq{
+    1: required i32 sourceDataNodeID
+    2: required i32 targetDataNodeID
+    3: required i32 schemaRegionID
+}
+
+struct MigrateDataRegionReq{
+    1: required i32 sourceDataNodeID
+    2: required i32 targetDataNodeID
+    3: required i32 dataRegionID
+}
 
 struct TFragmentInstanceId {
   1: required string queryId
@@ -113,6 +135,9 @@ struct SchemaFetchResponse {
 }
 
 service InternalService {
+
+    // -----------------------------------For Config Node-----------------------------------------------
+
     TSendFragmentInstanceResp sendFragmentInstance(TSendFragmentInstanceReq req);
 
     TFragmentInstanceStateResp fetchFragmentInstanceState(TFetchFragmentInstanceStateReq req);
@@ -124,6 +149,37 @@ service InternalService {
     TCancelResp cancelFragmentInstance(TCancelFragmentInstanceReq req);
 
     SchemaFetchResponse fetchSchema(SchemaFetchRequest req)
+
+
+    // -----------------------------------For Config Node-----------------------------------------------
+
+    /**
+      * Config node will create a schema region on a list of data nodes.
+      *
+      * @param data nodes of the schema region, and schema region id generated by config node
+    **/
+    common.TSStatus createSchemaRegion(CreateSchemaRegionReq req)
+
+    /**
+      * Config node will create a data region on a list of data nodes.
+      *
+      * @param data nodes of the data region, and data region id generated by config node
+    **/
+    common.TSStatus createDataRegion(CreateDataRegionReq req)
+
+    /**
+      * Config node will migrate a schema region from one data node to another
+      *
+      * @param previous data node in the schema region, new data node, and schema region id
+    **/
+    common.TSStatus migrateSchemaRegion(MigrateSchemaRegionReq req)
+
+    /**
+      * Config node will migrate a data region from one data node to another
+      *
+      * @param previous data node in the data region, new data node, and dataregion id
+    **/
+    common.TSStatus migrateDataRegion(MigrateDataRegionReq req)
 }
 
 service DataBlockService {
